[BE] UPDATE : #S08P12D104-172 엔티티(우산, 우산케이스) 컬럼 수정 & DB에 테스트 데이터 입력
Former-commit-id: 9504b46f0ef8e2a1abfd5084b94484a3374c2ec6

+++ backend/bpservice/src/main/java/kr/co/bpservice/entity/brolly/BrollyCase.java
@@ -17,7 +17,7 @@ public class BrollyCase {
     @Column(length = 50)
     private String name;
 
-    private Float lat;
+    private Double lat;
 
-    private Float lng;
+    private Double lng;
 }
